requirements.txt
beautifulsoup4==4.9.3
requests==2.24.0
pandas==1.1.3
transformers==4.38.0
pytorch==2.2.0